llm_bridge/requirements.txt
boto3 
python-dotenv 